feat: volume detection and notice
