+++ app/src/bcsc-theme/features/verify/live-call/LiveCallScreen.tsx
@@ -3,147 +3,21 @@ import { AppBannerSection as BannerSection } from '@/bcsc-theme/components/AppBa
 import useVideoCallFlow from '@/bcsc-theme/features/verify/live-call/hooks/useVideoCallFlow'
 import { VideoCallFlowState } from '@/bcsc-theme/features/verify/live-call/types/live-call'
 import { BCSCScreens, BCSCVerifyIdentityStackParams } from '@/bcsc-theme/types/navigators'
-import ProgressBar from '@/components/ProgressBar'
 import { BCDispatchAction, BCState } from '@/store'
-import Mountains from '@assets/img/mountains-circle.svg'
 import { Button, ButtonType, testIdWithKey, ThemedText, TOKENS, useServices, useStore, useTheme } from '@bifold/core'
 import { CommonActions } from '@react-navigation/native'
 import { StackNavigationProp } from '@react-navigation/stack'
 import { useCallback, useEffect, useMemo, useRef, useState } from 'react'
 import { useTranslation } from 'react-i18next'
-import { ColorValue, StyleSheet, TouchableOpacity, useWindowDimensions, View } from 'react-native'
+import { StyleSheet, useWindowDimensions, View } from 'react-native'
 import InCallManager from 'react-native-incall-manager'
 import { SafeAreaView } from 'react-native-safe-area-context'
 import Icon from 'react-native-vector-icons/MaterialCommunityIcons'
+import VolumeManager from 'react-native-volume-manager'
 import { MediaStreamTrack, RTCView } from 'react-native-webrtc'
+import CallIconButton from './components/CallIconButton'
+import CallLoadingView from './components/CallLoadingView'
 import { formatCallTime } from './utils/formatCallTime'
-
-type CallIconButtonProps = {
-  onPress: () => void
-  primaryColor: ColorValue
-  secondaryColor: ColorValue
-  size: number
-  iconName: string
-  label: string
-  testIDKey: string
-}
-
-const CallIconButton = ({
-  onPress,
-  primaryColor,
-  secondaryColor,
-  size,
-  iconName,
-  label,
-  testIDKey,
-}: CallIconButtonProps) => {
-  const { Spacing } = useTheme()
-  const styles = StyleSheet.create({
-    iconButton: {
-      width: size,
-      height: size,
-      borderRadius: size / 2,
-      justifyContent: 'center',
-      alignItems: 'center',
-      backgroundColor: secondaryColor,
-      borderWidth: 1,
-      borderColor: primaryColor,
-    },
-  })
-
-  return (
-    <TouchableOpacity
-      onPress={onPress}
-      testID={testIdWithKey(testIDKey)}
-      accessibilityLabel={label}
-      style={{ justifyContent: 'center', alignItems: 'center' }}
-    >
-      <View style={styles.iconButton}>
-        <Icon name={iconName} size={size - 32} color={primaryColor} />
-      </View>
-      <ThemedText style={{ color: primaryColor, marginTop: Spacing.xs }}>{label}</ThemedText>
-    </TouchableOpacity>
-  )
-}
-
-type LoadingViewProps = {
-  onCancel: () => void
-  message?: string
-}
-
-const LoadingView = ({ onCancel, message }: LoadingViewProps) => {
-  const { Spacing, ColorPalette } = useTheme()
-  const { t } = useTranslation()
-  const [progressPercent, setProgressPercent] = useState(0)
-  const [delayReached, setDelayReached] = useState(false)
-
-  const styles = StyleSheet.create({
-    pageContainer: {
-      flex: 1,
-      justifyContent: 'space-between',
-      backgroundColor: ColorPalette.brand.primaryBackground,
-      padding: Spacing.md,
-    },
-    contentContainer: {
-      flex: 1,
-    },
-    controlsContainer: {
-      marginTop: 'auto',
-    },
-  })
-
-  useEffect(() => {
-    const interval = setInterval(() => {
-      setProgressPercent((prev) => {
-        if (prev > 80) {
-          setDelayReached(true)
-        }
-
-        // Logarithmic progression - slows down as it approaches 100%
-        const maxProgress = 97
-
-        if (prev >= maxProgress) {
-          return prev
-        }
-
-        const increment = (maxProgress - prev) * 0.02
-
-        return Math.min(prev + increment, maxProgress)
-      })
-    }, 100)
-
-    return () => clearInterval(interval)
-  }, [])
-
-  return (
-    <SafeAreaView style={styles.pageContainer}>
-      <View style={styles.contentContainer}>
-        <ProgressBar dark={true} progressPercent={progressPercent} />
-        <ThemedText variant={'headingTwo'} style={{ marginTop: 2 * Spacing.xxl, textAlign: 'center' }}>
-          {t('Unified.VideoCall.OneMomentPlease')}
-        </ThemedText>
-        <ThemedText style={{ marginTop: 2 * Spacing.xxl, textAlign: 'center' }}>
-          {message || t('Unified.VideoCall.SettingThingsUp')}
-        </ThemedText>
-        <Mountains style={{ alignSelf: 'center', marginVertical: Spacing.md }} height={200} width={200} />
-        {delayReached ? (
-          <ThemedText variant={'labelSubtitle'} style={{ textAlign: 'center' }}>
-            {t('Unified.VideoCall.TakingLongerThanUsual')}
-          </ThemedText>
-        ) : null}
-      </View>
-      <View style={styles.controlsContainer}>
-        <Button
-          buttonType={ButtonType.Primary}
-          onPress={onCancel}
-          title={t('Global.Cancel')}
-          accessibilityLabel={t('Global.Cancel')}
-          testID={testIdWithKey('Cancel')}
-        />
-      </View>
-    </SafeAreaView>
-  )
-}
 
 type LiveCallScreenProps = {
   navigation: StackNavigationProp<BCSCVerifyIdentityStackParams, BCSCScreens.LiveCall>
@@ -159,6 +33,7 @@ const LiveCallScreen = ({ navigation }: LiveCallScreenProps) => {
   const [onMute, setOnMute] = useState(false)
   const [callStartTime, setCallStartTime] = useState<number | null>(null)
   const [callTimer, setCallTimer] = useState<string>('')
+  const [systemVolume, setSystemVolume] = useState<number>(1)
   const timerIntervalRef = useRef<NodeJS.Timeout | null>(null)
   const { token } = useApi()
   const [logger] = useServices([TOKENS.UTIL_LOGGER])
@@ -217,10 +92,11 @@ const LiveCallScreen = ({ navigation }: LiveCallScreenProps) => {
     videoCallError,
     localStream,
     remoteStream,
+    isInBackground,
     startVideoCall,
     cleanup,
     retryConnection,
-    isInBackground,
+    setCallEnded,
   } = useVideoCallFlow(leaveCall)
 
   const inCall = useMemo(() => flowState === VideoCallFlowState.IN_CALL, [flowState])
@@ -270,6 +146,28 @@ const LiveCallScreen = ({ navigation }: LiveCallScreenProps) => {
     }
   }, [])
 
+  // Monitor device volume changes
+  useEffect(() => {
+    const getInitialVolume = async () => {
+      try {
+        const volume = await VolumeManager.getVolume()
+        setSystemVolume(volume.volume)
+      } catch (error) {
+        logger.warn('Failed to get initial volume', { error: error as Error })
+      }
+    }
+
+    const volumeListener = VolumeManager.addVolumeListener((result) => {
+      setSystemVolume(result.volume)
+    })
+
+    getInitialVolume()
+
+    return () => {
+      volumeListener?.remove()
+    }
+  }, [logger])
+
   const banner: { type: 'warning' | 'error'; title: string } | null = useMemo(() => {
     if (isInBackground) {
       return { type: 'warning', title: t('Unified.VideoCall.VideoWillResume') }
@@ -277,12 +175,18 @@ const LiveCallScreen = ({ navigation }: LiveCallScreenProps) => {
     if (videoHidden) {
       return { type: 'error', title: t('Unified.VideoCall.AgentCantSeeYou') }
     }
+    if (systemVolume === 0) {
+      return { type: 'error', title: 'Please turn up the volume to hear the agent' }
+    }
     if (onMute) {
       return { type: 'error', title: t('Unified.VideoCall.AgentCantHearYou') }
     }
+    if (systemVolume < 0.2) {
+      return { type: 'warning', title: 'Your volume is low, you may need to turn it up to hear the agent' }
+    }
 
     return null
-  }, [isInBackground, onMute, videoHidden, t])
+  }, [isInBackground, onMute, videoHidden, inCall, systemVolume, t])
 
   useEffect(() => {
     if (localStream) {
@@ -338,6 +242,7 @@ const LiveCallScreen = ({ navigation }: LiveCallScreenProps) => {
   const handleEndCall = useCallback(async () => {
     try {
       await cleanup()
+      setCallEnded()
       await leaveCall()
     } catch (error) {
       logger.error('Error while leaving video call', error as Error)
@@ -376,8 +281,12 @@ const LiveCallScreen = ({ navigation }: LiveCallScreenProps) => {
     )
   }
 
+  if (flowState === VideoCallFlowState.CALL_ENDED) {
+    return <CallLoadingView message={t('Unified.VideoCall.CallStates.CallEnded')} />
+  }
+
   if (flowState !== VideoCallFlowState.IN_CALL) {
-    return <LoadingView onCancel={handleEndCall} message={stateMessage || undefined} />
+    return <CallLoadingView onCancel={handleEndCall} message={stateMessage || undefined} />
   }
 
   const styles = StyleSheet.create({
@@ -392,8 +301,7 @@ const LiveCallScreen = ({ navigation }: LiveCallScreenProps) => {
       right: 0,
       bottom: '15%',
       flex: 1,
-      // zoom
-      transform: [{ scale: 1.5 }],
+      transform: [{ scale: 1.5 }], // zoom
     },
     // just helpful labels, no properties needed
     upperContainer: {},
