+++ app/src/localization/fr/index.ts
@@ -336,7 +336,8 @@ const translation = {
         "CreatingSession": "Creating video session... (FR)",
         "ConnectingWebRTC": "Connecting to video service... (FR)",
         "WaitingForAgent": "Waiting for an agent to join... (FR)",
-        "Initializing": "Initializing... (FR)"
+        "Initializing": "Initializing... (FR)",
+        "CallEnded": "Call ended, checking verification... (FR)"
       },
       "Loading": {
         "OneMomentPlease": "One moment please... (FR)",
@@ -347,7 +348,9 @@ const translation = {
       "Banners": {
         "VideoWillResume": "Video will resume when you return to this app (FR)",
         "AgentCantSeeYou": "Agent can't see you while your video is off (FR)",
-        "AgentCantHearYou": "Agent can't hear you while your microphone is muted (FR)"
+        "AgentCantHearYou": "Agent can't hear you while your microphone is muted (FR)",
+        "VolumeOff": "Please turn up your volume to hear the agent (FR)",
+        "VolumeLow": "Your volume is low, you may need to turn it up to hear the agent (FR)"
       },
       "Errors": {
         "ConnectionError": "Connection Error (FR)",
